ISSUE #382 fix compilation

+++ test/src/unit/repo/manipulator/modeloptimizer/ut_repo_optimizer_multipart.cpp
@@ -55,7 +55,7 @@ repo::core::model::MeshNode* createRandomMesh(const bool hasUV, const bool isInd
 	std::vector<repo_face_t> faces;
 
 	for (int i = 0; i < nVertices; ++i) {
-		vertices.push_back({ std::rand(), std::rand(), std::rand() });
+		vertices.push_back({ (float)std::rand(), (float)std::rand(), (float)std::rand() });
 	}
 
 	for (int i = 0; i < nFaces; ++i) {
@@ -74,7 +74,7 @@ repo::core::model::MeshNode* createRandomMesh(const bool hasUV, const bool isInd
 		uvs.push_back(channel);
 	}
 
-	auto mesh = new repo::core::model::MeshNode(repo::core::model::RepoBSONFactory::makeMeshNode(vertices, faces, {}, {}, uvs, {}, {}, "mesh",parent));
+	auto mesh = new repo::core::model::MeshNode(repo::core::model::RepoBSONFactory::makeMeshNode(vertices, faces, {}, {}, uvs, {}, {}, "mesh", parent));
 
 	if (isIndependent)
 		mesh->swap(mesh->cloneAndFlagIndependent());
@@ -91,19 +91,18 @@ TEST(MultipartOptimizer, TestAllMerged)
 	auto nMesh = 3;
 	repo::core::model::RepoNodeSet meshes, trans, dummy;
 	trans.insert(root);
-	for(int i =0; i < nMesh; ++i)
-		meshes.insert(createRandomMesh(false, false, {rootID}));
+	for (int i = 0; i < nMesh; ++i)
+		meshes.insert(createRandomMesh(false, false, { rootID }));
 
 	repo::core::model::RepoScene *scene = new repo::core::model::RepoScene({}, dummy, meshes, dummy, dummy, dummy, trans);
 	ASSERT_TRUE(scene->hasRoot(DEFAULT_GRAPH));
 	ASSERT_FALSE(scene->hasRoot(OPTIMIZED_GRAPH));
-	
+
 	EXPECT_TRUE(opt.apply(scene));
 	EXPECT_TRUE(scene->hasRoot(DEFAULT_GRAPH));
 	EXPECT_TRUE(scene->hasRoot(OPTIMIZED_GRAPH));
 
 	EXPECT_EQ(1, scene->getAllMeshes(OPTIMIZED_GRAPH).size());
-
 }
 
 TEST(MultipartOptimizer, TestWithIndependent)
@@ -127,7 +126,6 @@ TEST(MultipartOptimizer, TestWithIndependent)
 	EXPECT_TRUE(scene->hasRoot(OPTIMIZED_GRAPH));
 
 	EXPECT_EQ(2, scene->getAllMeshes(OPTIMIZED_GRAPH).size());
-
 }
 
 TEST(MultipartOptimizer, TestWithUV)
@@ -151,5 +149,4 @@ TEST(MultipartOptimizer, TestWithUV)
 	EXPECT_TRUE(scene->hasRoot(OPTIMIZED_GRAPH));
 
 	EXPECT_EQ(2, scene->getAllMeshes(OPTIMIZED_GRAPH).size());
-
 }
